requirements.txt
# langchain
# python-dotenv
# ipykernel
# langchain-community
# pypdf
# bs4
# arxiv
# pymupdf
# wikipedia
# langchain-text-splitters
# langchain-openai
# chromadb
# sentence_transformers
# langchain_huggingface
# faiss-cpu
# langchain_chroma
# duckdb
# pandas
# openai
# langchain-groq
# duckduckgo_search==5.3.1b1
# pymupdf
# arxiv
# wikipedia
# mysql-connector-python
# SQLAlchemy
# validators==0.28.1
# youtube_transcript_api
# unstructured
# pytube
# numexpr
# huggingface_hub
# youtube_transcript_api==0.6.2
# streamlit
# google-generativeai
# requests
# beautifulsoup4
# newspaper3k
# lxml_html_clean
# google-generativeai


streamlit
google-generativeai==0.7.2  # Replace with the actual version if available
youtube-transcript-api==0.6.0
requests==2.28.2
beautifulsoup4==4.12.2
